requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.10.1
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
coloredlogs==15.0.1
conda==23.1.0
conda-package-handling==1.8.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.5
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.4.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
humanfriendly==10.0
idna==3.4
imageio==2.31.6
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning==2.1.1
lightning-utilities==0.9.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.26.1
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.1
onnxsim==0.4.35
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openvino==2023.1.0
openvino-telemetry==2023.2.1
packaging==23.2
pandas==2.1.2
pathspec==0.11.2
Pillow==9.3.0
platformdirs==3.11.0
protobuf==4.23.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
Pygments==2.16.1
pylance==0.8.7
pyparsing==3.1.1
pyqtgraph==0.13.3
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.8.2
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
shiboken6==6.6.0
six==1.16.0
sympy==1.12
tensorboard==2.15.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
torchmetrics==1.2.0
torchvision==0.16.0+cpu
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.202
urllib3==1.26.13
Werkzeug==3.0.0
yarl==1.9.2
